Changes
Updated how the user gets retrieved, now from vuex store.
Updated the unique ID in the bros data in firebase to be the email before @ sign instead of some gibberish UID that I can't easily decipher
Added that UID to the store so that I don't have to do the substring to get the email before the @ sign everytime
Added brother vouching to front door mode, with autocomplete!
Added ability to remove guests you added
Made it so that if an admin deleted your account you get notified that you have to finish the deletion
Fixed a bug with making an account
Added a list of brothers names to be populated in the brother vouching input, will auto update everytime someone opens the social admin menu